Extract `Plan` translation tests to a separate file
