+++ test/execution/test_plan.py
@@ -8,11 +8,9 @@ import torch.optim as optim
 
 import syft as sy
 from itertools import starmap
-from syft.generic.pointers.pointer_tensor import PointerTensor
 from syft.generic.frameworks.types import FrameworkTensor
 from syft.execution.placeholder import PlaceHolder
 from syft.execution.plan import Plan
-from syft.execution.translation.torchscript import PlanTranslatorTorchscript
 from syft.serde.serde import deserialize
 from syft.serde.serde import serialize
 
@@ -881,26 +879,6 @@ def test_train_plan_locally_and_then_send_it(hook, start_remote_worker):
     server.terminate()
 
 
-# Plans are not supporting non-local computation flows for the moment
-# def test_send_with_plan(workers):
-#     bob = workers["bob"]
-#
-#     raise NotImplementedError
-#
-#     @sy.func2plan([th.Size((1, 3))])
-#     def plan_double_abs(x):
-#         x = x.send(bob)
-#         x = x + x
-#         x = th.abs(x)
-#         return x
-#
-#     expected = th.tensor([4.0, 4.0, 4.0])
-#     ptr_result = plan_double_abs(th.tensor([-2.0, 2.0, 2.0]))
-#     assert isinstance(ptr_result.child, sy.PointerTensor)
-#     result = ptr_result.get()
-#     assert th.equal(result, expected)
-
-
 def test_cached_plan_send(workers):
     bob = workers["bob"]
 
@@ -1136,119 +1114,3 @@ def test_plan_input_usage(hook):
     pointer_to_result = pointer_plan(pointer_to_data_1, pointer_to_data_2)
     result = pointer_to_result.get()
     assert (result == x12).all
-
-
-def test_func_plan_can_be_translated_to_torchscript(hook, workers):
-    # Disable build time auto translation
-    Plan._build_translators = []
-
-    @sy.func2plan(args_shape=[(3, 3)])
-    def plan(x):
-        x = x * 2
-        x = x.abs()
-        return x
-
-    orig_plan = plan.copy()
-
-    inp = th.tensor([1, -1, 2])
-    res1 = plan(inp)
-    plan.add_translation(PlanTranslatorTorchscript)
-    res2 = plan.torchscript(inp)
-    assert (res1 == res2).all()
-
-    # check that translation can be done after serde
-    serde_plan = deserialize(serialize(orig_plan))
-    serde_plan.add_translation(PlanTranslatorTorchscript)
-    res3 = serde_plan.torchscript(inp)
-    assert (res1 == res3).all()
-
-    # check that translation is not lost after serde
-    serde_plan_full = deserialize(serialize(plan))
-    res4 = serde_plan_full.torchscript(inp)
-    assert (res1 == res4).all()
-
-
-def test_cls_plan_can_be_translated_to_torchscript(hook, workers):
-    # Disable build time auto translation
-    Plan._build_translators = []
-
-    class Net(sy.Plan):
-        def __init__(self):
-            super(Net, self).__init__()
-            self.fc1 = nn.Linear(2, 3)
-            self.fc2 = nn.Linear(3, 1)
-
-        def forward(self, x):
-            x = self.fc1(x)
-            x = F.relu(x)
-            x = self.fc2(x)
-            return x
-
-    plan = Net()
-    plan.build(th.zeros(10, 2))
-    orig_plan = plan.copy()
-
-    inp = th.randn(10, 2)
-
-    res1 = plan(inp)
-    plan.add_translation(PlanTranslatorTorchscript)
-    res2 = plan.torchscript(inp, plan.parameters())
-    assert (res1 == res2).all()
-
-    # check that translation can be done after serde
-    serde_plan = deserialize(serialize(orig_plan))
-    serde_plan.add_translation(PlanTranslatorTorchscript)
-    res3 = serde_plan.torchscript(inp, serde_plan.parameters())
-    assert (res1 == res3).all()
-
-    # check that translation is not lost after serde
-    serde_plan_full = deserialize(serialize(plan))
-    res4 = serde_plan_full.torchscript(inp, serde_plan_full.parameters())
-    assert (res1 == res4).all()
-
-
-def test_plan_translation_remove(hook, workers):
-    # Disable build time auto translation
-    Plan._build_translators = []
-
-    @sy.func2plan(args_shape=[(3, 3)])
-    def plan(x):
-        x = x * 2
-        x = x.abs()
-        return x
-
-    plan.add_translation(PlanTranslatorTorchscript)
-
-    full_plan = plan.copy()
-    assert full_plan.torchscript is not None
-
-    assert plan.torchscript is not None
-    assert len(plan.role.actions) > 0
-
-    plan.remove_translation()
-    assert plan.torchscript is not None
-    assert len(plan.role.actions) == 0
-
-    plan.remove_translation(PlanTranslatorTorchscript)
-    assert plan.torchscript is None
-    assert len(plan.role.actions) == 0
-
-    full_plan.remove_translation(PlanTranslatorTorchscript)
-    assert full_plan.torchscript is None
-    assert len(full_plan.role.actions) > 0
-
-
-def test_plan_translated_on_build(hook, workers):
-    # Enable torchscript translator
-    Plan.register_build_translator(PlanTranslatorTorchscript)
-
-    @sy.func2plan(args_shape=[(3, 3)])
-    def plan(x):
-        x = x * 2
-        x = x.abs()
-        return x
-
-    inp = th.tensor([1, -1, 2])
-    res1 = plan(inp)
-    res2 = plan.torchscript(inp)
-    assert (res1 == res2).all()
